third_party/blink/web_tests/platform/win/external/wpt/paint-timing/fcp-only/fcp-ensure-update-the-rendering-step-expected.txt

This is a testharness.js-based test.
FAIL The first-contentful-paint timestamp should be same as the last RAF assert_approx_equals: One of the past requestAnimationFrame should have the same timestamp as paint entry expected 49.281 +/- 1 but got 55.11999999907857

assert_approx_equals: One of the past requestAnimationFrame should have the same timestamp as paint entry expected 40.982 +/- 1 but got 55.11999999907857
Harness: the test ran to completion.
